### 2023-03-06 #199 @whedon
[PRE REVIEW]: Manim Slides: A Python package for presenting Manim content anywhere
**Submitting author:** @jeertmans (<a href="http://orcid.org/0000-0002-5579-5360">Jérome Eertmans</a>)
**Repository:** <a href="https://github.com/jeertmans/manim-slides" target ="_blank">https://github.com/jeertmans/manim-slides</a>
**Version:** v4.9.2
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
