### 2023-03-08 9bb691f
chore(lib): use `next_slide` not `pause`
This deprecates `pause` function in favor to `next_slide`, that will also be called by `next_section` in the future.
