+++ pyproject.toml
@@ -71,10 +71,10 @@ opencv-python = "^4.6.0.66"
 pydantic = "^1.10.2"
 pyside6 = "^6.4.1"
 python = ">=3.8.1,<3.12"
+python-pptx = "^0.6.21"
 requests = "^2.28.1"
 rich = "^13.3.2"
 tqdm = "^4.64.1"
-python-pptx = "^0.6.21"
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.10.0"
