### 2023-03-08 68d8aa4
chore(deps): remove pkg_resources in favor to importlib
This is what pkg_resources' team recommends
