+++ manim_slides/convert.py
@@ -23,7 +23,7 @@ def open_with_default(file: Path):
     if system == "Darwin":
         subprocess.call(("open", str(file)))
     elif system == "Windows":
-        os.startfile(str(file))
+        os.startfile(str(file))  # type: ignore[attr-defined]
     else:
         subprocess.call(("xdg-open", str(file)))
 
