+++ pyproject.toml
@@ -72,6 +72,7 @@ pydantic = "^1.10.2"
 pyside6 = "^6.4.1"
 python = ">=3.8.1,<3.12"
 requests = "^2.28.1"
+rich = "^13.3.2"
 tqdm = "^4.64.1"
 
 [tool.poetry.group.dev.dependencies]
