+++ pyproject.toml
@@ -66,6 +66,7 @@ version = "4.9.2"
 [tool.poetry.dependencies]
 click = "^8.1.3"
 click-default-group = "^1.2.2"
+lxml = "^4.9.2"
 numpy = "^1.19"
 opencv-python = "^4.6.0.66"
 pydantic = "^1.10.2"
@@ -75,7 +76,6 @@ python-pptx = "^0.6.21"
 requests = "^2.28.1"
 rich = "^13.3.2"
 tqdm = "^4.64.1"
-lxml = "^4.9.2"
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.10.0"
