### 2023-03-09 c384e9a
fix(deps): if case for Python 3.8 (<3.9)
