### 2023-03-09 4734be5
chore(docs): document new feature
