+++ example.py
@@ -22,14 +22,14 @@ class BasicExample(Slide):
         dot = Dot()
 
         self.play(GrowFromCenter(circle))
-        self.pause()  # Waits user to press continue to go to the next slide
+        self.next_slide()  # Waits user to press continue to go to the next slide
 
         self.start_loop()  # Start loop
         self.play(MoveAlongPath(dot, circle), run_time=2, rate_func=linear)
         self.end_loop()  # This will loop until user inputs a key
 
         self.play(dot.animate.move_to(ORIGIN))
-        self.pause()  # Waits user to press continue to go to the next slide
+        self.next_slide()  # Waits user to press continue to go to the next slide
 
 
 class TestFileTooLong(Slide):
@@ -47,7 +47,7 @@ class TestFileTooLong(Slide):
             self.play(dot.animate.move_to(direction), run_time=0.1)
             self.play(dot.animate.move_to(ORIGIN), run_time=0.1)
 
-        self.pause()
+        self.next_slide()
 
 
 class ConvertExample(Slide):
@@ -77,7 +77,7 @@ class ConvertExample(Slide):
 
         self.play(FadeIn(title))
 
-        self.pause()
+        self.next_slide()
 
         code = Code(
             code="""from manim import *
@@ -146,10 +146,10 @@ class Example(Slide):
         self.add(dot)
 
         self.play(Indicate(dot, scale_factor=2))
-        self.pause()
+        self.next_slide()
         square = Square()
         self.play(Transform(dot, square))
-        self.pause()
+        self.next_slide()
         self.play(Rotate(square, angle=PI/2))
 """,
             language="python",
@@ -168,7 +168,7 @@ class Example(Slide):
         self.end_loop()
         square = Square()
         self.play(Transform(dot, square))
-        self.pause()
+        self.next_slide()
         self.play(Rotate(square, angle=PI/2))
 """,
             language="python",
@@ -195,38 +195,38 @@ class Example(Slide):
 
         self.play(FadeIn(code))
         self.tinywait()
-        self.pause()
+        self.next_slide()
 
         self.play(FadeIn(step, shift=RIGHT))
         self.play(Transform(code, code_step_1))
         self.tinywait()
-        self.pause()
+        self.next_slide()
 
         self.play(Transform(step, step_2))
         self.play(Transform(code, code_step_2))
         self.tinywait()
-        self.pause()
+        self.next_slide()
 
         self.play(Transform(step, step_3))
         self.play(Transform(code, code_step_3))
         self.tinywait()
-        self.pause()
+        self.next_slide()
 
         self.play(Transform(step, step_4))
         self.play(Transform(code, code_step_4))
         self.tinywait()
-        self.pause()
+        self.next_slide()
 
         self.play(Transform(step, step_5))
         self.play(Transform(code, code_step_5))
         self.tinywait()
-        self.pause()
+        self.next_slide()
 
         self.play(Transform(step, step_6))
         self.play(Transform(code, code_step_6))
         self.play(code.animate.shift(UP), FadeIn(code_step_7), FadeIn(or_text))
         self.tinywait()
-        self.pause()
+        self.next_slide()
 
         watch_text = Text("Watch result on next slides!").shift(2 * DOWN).scale(0.5)
 
@@ -246,10 +246,10 @@ class Example(Slide):
         self.remove(dot)
         self.add(square)
         self.tinywait()
-        self.pause()
+        self.next_slide()
         self.play(Rotate(square, angle=PI / 4))
         self.tinywait()
-        self.pause()
+        self.next_slide()
 
         learn_more_text = (
             VGroup(
@@ -281,7 +281,7 @@ if not MANIMGL:
[…]
