+++ manim_slides/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "4.9.2"
+__version__ = "4.10.0"
