+++ pyproject.toml
@@ -61,7 +61,7 @@ packages = [
 ]
 readme = "README.md"
 repository = "https://github.com/jeertmans/manim-slides"
-version = "4.9.2"
+version = "4.10.0"
 
 [tool.poetry.dependencies]
 click = "^8.1.3"
