+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 4.9.2
+current_version = 4.10.0
 commit = True
 message = chore(version): bump {current_version} to {new_version}
 
