### 2023-03-09 149b12f
chore(lib): raise error if skipping
Temporary error before implementing this
