### 2023-03-10 ab10e00
feat(render): support skipping animations
Manim Slides now supports when Manim or ManimGL skip rendering some animations. All non-public attributes or methods are now named with leading __
