+++ manim_slides/slide.py
@@ -47,15 +47,15 @@ class Slide(Scene):  # type:ignore
 
         super().__init__(*args, **kwargs)
 
-        self.output_folder = output_folder
-        self.slides: List[SlideConfig] = []
-        self.current_slide = 1
-        self.current_animation = 0
-        self.loop_start_animation: Optional[int] = None
-        self.pause_start_animation = 0
+        self.__output_folder = output_folder
+        self.__slides: List[SlideConfig] = []
+        self.__current_slide = 1
+        self.__current_animation = 0
+        self.__loop_start_animation: Optional[int] = None
+        self.__pause_start_animation = 0
 
     @property
-    def partial_movie_files(self) -> List[str]:
+    def __partial_movie_files(self) -> List[str]:
         """Returns a list of partial movie files, a.k.a animations."""
         if MANIMGL:
             from manimlib.utils.file_ops import get_sorted_integer_files
@@ -71,25 +71,32 @@ class Slide(Scene):  # type:ignore
             return self.renderer.file_writer.partial_movie_files
 
     @property
-    def show_progress_bar(self) -> bool:
+    def __show_progress_bar(self) -> bool:
         """Returns True if progress bar should be displayed."""
         if MANIMGL:
-            return getattr(super(Scene, self), "show_progress_bar", True)
+            return getattr(self, "show_progress_bar", True)
         else:
             return config["progress_bar"] != "none"
 
     @property
-    def leave_progress_bar(self) -> bool:
+    def __leave_progress_bar(self) -> bool:
         """Returns True if progress bar should be left after completed."""
         if MANIMGL:
-            return getattr(super(Scene, self), "leave_progress_bars", False)
+            return getattr(self, "leave_progress_bars", False)
         else:
             return config["progress_bar"] == "leave"
+
+    @property
+    def __start_at_animation_number(self) -> Optional[int]:
+        if MANIMGL:
+            return getattr(self, "start_at_animation_number", None)
+        else:
+            return config["from_animation_number"]
 
     def play(self, *args: Any, **kwargs: Any) -> None:
         """Overloads `self.play` and increment animation count."""
         super().play(*args, **kwargs)
-        self.current_animation += 1
+        self.__current_animation += 1
 
     def next_slide(self):
         """
@@ -133,19 +140,19 @@ class Slide(Scene):  # type:ignore
                     self.play(FadeOut(text))
         """
         assert (
-            self.loop_start_animation is None
+            self.__loop_start_animation is None
         ), "You cannot call `self.next_slide()` inside a loop"
 
-        self.slides.append(
+        self.__slides.append(
             SlideConfig(
                 type=SlideType.slide,
-                start_animation=self.pause_start_animation,
-                end_animation=self.current_animation,
-                number=self.current_slide,
+                start_animation=self.__pause_start_animation,
+                end_animation=self.__current_animation,
+                number=self.__current_slide,
             )
         )
-        self.current_slide += 1
-        self.pause_start_animation = self.current_animation
+        self.__current_slide += 1
+        self.__pause_start_animation = self.__current_animation
 
     def pause(self) -> None:
         """
@@ -161,22 +168,22 @@ class Slide(Scene):  # type:ignore
         )
         self.next_slide()
 
-    def add_last_slide(self) -> None:
+    def __add_last_slide(self) -> None:
         """Adds a 'last' slide to the end of slides."""
 
         if (
-            len(self.slides) > 0
-            and self.current_animation == self.slides[-1].end_animation
+            len(self.__slides) > 0
+            and self.__current_animation == self.__slides[-1].end_animation
         ):
-            self.slides[-1].type = SlideType.last
+            self.__slides[-1].type = SlideType.last
             return
 
-        self.slides.append(
[…]
