### 2023-03-11 ba9a9f0
feat(cli): add start at index options
