+++ manim_slides/present.py
@@ -73,7 +73,7 @@ class Presentation:
         self.slides: List[SlideConfig] = config.slides
         self.files: List[str] = config.files
 
-        self.current_slide_index: int = 0
+        self.__current_slide_index: int = 0
         self.current_animation: int = self.current_slide.start_animation
         self.current_file: str = ""
 
@@ -87,10 +87,45 @@ class Presentation:
 
         self.reset()
 
-    def animation_is_in_current_slide(self, animation: int) -> bool:
-        """Returns true if animation is played by the current slide."""
-        slide = self.current_slide
-        return slide.start_animation <= animation < slide.end_animation
+    def __len__(self) -> int:
+        return len(self.slides)
+
+    @property
+    def current_slide_index(self) -> int:
+        return self.__current_slide_index
+
+    @current_slide_index.setter
+    def current_slide_index(self, value: Optional[int]):
+        if value:
+            if -len(self) <= value < len(self):
+                self.__current_slide_index = value
+                self.current_animation = self.current_slide.start_animation
+            else:
+                logger.error(
+                    f"Could not load slide number {value}, playing first slide instead."
+                )
+
+    def set_current_animation_and_update_slide_number(self, value: Optional[int]):
+        if value:
+            n_files = len(self.files)
+            if -n_files <= value < n_files:
+                if value < 0:
+                    value += n_files
+
+                for i, slide in enumerate(self.slides):
+                    if value < slide.end_animation:
+                        self.current_slide_index = i
+                        self.current_animation = value
+                        return
+
+                assert (
+                    False
+                ), f"An error occurred when setting the current animation to {value}, please create an issue on GitHub!"
+
+            else:
+                logger.error(
+                    f"Could not load animation number {value}, playing first animation instead."
+                )
 
     @property
     def current_slide(self) -> SlideConfig:
@@ -295,9 +330,9 @@ class Display(QThread):  # type: ignore
         skip_all: bool = False,
         record_to: Optional[str] = None,
         exit_after_last_slide: bool = False,
-        start_at_scene_number: int = 0,
-        start_at_slide_number: int = 0,
-        start_at_animation_number: int = 0,
+        start_at_scene_number: Optional[int] = None,
+        start_at_slide_number: Optional[int] = None,
+        start_at_animation_number: Optional[int] = None,
     ) -> None:
         super().__init__()
         self.presentations = presentations
@@ -309,31 +344,33 @@ class Display(QThread):  # type: ignore
 
         self.state = State.PLAYING
         self.lastframe: Optional[np.ndarray] = None
-        try:
-            self.current_presentation_index = start_at_scene_number
-            presentation = self.current_presentation
 
-            old_slide_index = presentation.current_slide_index
-            try:
-                presentation.current_slide_index = start_at_slide_number
-                _ = presentation.current_slide
-            except IndexError:
-                logger.error(f"Could not load slide number {start_at_slide_number}, playing slide number {old_slide_index} instead.")
-                presentation.current_slide_index = old_slide_index
-
-            if presentation.animation_is_in_current_slide(start_at_animation_number):
-                presentation.current_animation = start_at_animation_number
-            else:
-                logger.error(f"Could not load animation number {start_at_animation_number}, playing first animation number instead.")
-
-        except IndexError:
-            logger.error(f"Could not load scene number {start_at_scene_number}, playing first scene instead.")
-            self.current_presentation_index = 0
[…]
