### 2023-03-16 a38c32b
fix(lib): better to test if not None
