+++ manim_slides/present.py
@@ -9,6 +9,7 @@ from typing import Any, Dict, List, Optional, Tuple, Union
 import click
 import cv2
 import numpy as np
+from click import Context, Parameter
 from pydantic import ValidationError
 from PySide6.QtCore import Qt, QThread, Signal, Slot
 from PySide6.QtGui import QCloseEvent, QIcon, QImage, QKeyEvent, QPixmap, QResizeEvent
@@ -70,8 +71,7 @@ class Presentation:
     """Creates presentation from a configuration object."""
 
     def __init__(self, config: PresentationConfig) -> None:
-        self.slides: List[SlideConfig] = config.slides
-        self.files: List[str] = config.files
+        self.config = config
 
         self.__current_slide_index: int = 0
         self.current_animation: int = self.current_slide.start_animation
@@ -91,22 +91,40 @@ class Presentation:
         return len(self.slides)
 
     @property
+    def slides(self) -> List[SlideConfig]:
+        """Returns the list of slides."""
+        return self.config.slides
+
+    @property
+    def files(self) -> List[Path]:
+        """Returns the list of animation files."""
+        return self.config.files
+
+    @property
+    def resolution(self) -> Tuple[int, int]:
+        """Returns the resolution."""
+        return self.config.resolution
+
+    @property
     def current_slide_index(self) -> int:
         return self.__current_slide_index
 
     @current_slide_index.setter
-    def current_slide_index(self, value: Optional[int]):
-        if value:
+    def current_slide_index(self, value: Optional[int]) -> None:
+        if value is not None:
             if -len(self) <= value < len(self):
                 self.__current_slide_index = value
                 self.current_animation = self.current_slide.start_animation
+                logger.debug(f"Set current slide index to {value}")
             else:
                 logger.error(
                     f"Could not load slide number {value}, playing first slide instead."
                 )
 
-    def set_current_animation_and_update_slide_number(self, value: Optional[int]):
-        if value:
+    def set_current_animation_and_update_slide_number(
+        self, value: Optional[int]
+    ) -> None:
+        if value is not None:
             n_files = len(self.files)
             if -n_files <= value < n_files:
                 if value < 0:
@@ -116,6 +134,8 @@ class Presentation:
                     if value < slide.end_animation:
                         self.current_slide_index = i
                         self.current_animation = value
+
+                        logger.debug(f"Playing animation {value}, at slide index {i}")
                         return
 
                 assert (
@@ -159,7 +179,7 @@ class Presentation:
 
             self.release_cap()
 
-            file: str = self.files[animation]
+            file: str = str(self.files[animation])
 
             if self.reverse:
                 file = "{}_reversed{}".format(*os.path.splitext(file))
@@ -178,7 +198,7 @@ class Presentation:
 
     def rewind_current_slide(self) -> None:
         """Rewinds current slide to first frame."""
-        logger.debug("Rewinding curring slide")
+        logger.debug("Rewinding current slide")
         if self.reverse:
             self.current_animation = self.current_slide.end_animation - 1
         else:
@@ -216,9 +236,10 @@ class Presentation:
 
     def load_previous_slide(self) -> None:
         """Loads previous slide."""
-        logger.debug("Loading previous slide")
+        logger.debug(f"Loading previous slide, current is {self.current_slide_index}")
         self.cancel_reverse()
         self.current_slide_index = max(0, self.current_slide_index - 1)
+        logger.debug(f"Loading slide index {self.current_slide_index}")
         self.rewind_current_slide()
 
[…]
