### 2023-03-16 bec16f2
Use scene background color as default
