### 2023-03-20 712023d
fix(cli): do not terminate slides early
When a slide is replayed (either normally or reversed), its state must be reset.

Closes #161
