### 2023-03-20 2ee99da
fix(lib): prevent calling child method
Change `pause` to not call eventual child's `next_slide` method
