### 2023-03-22 4b4b886
Minor changes to feature: Read scene background
