### 2023-03-22 c4ca277
Small fix for feature "Read bg color"
