### 2023-03-22 3c6cf94
chore(ci): add typing ignore
