### 2023-03-23 #199 @labarba
@magsol Have a look at this submission 👀
https://eertmans.be/manim-slides/

It makes slides using the `manim` Python package. Cool!
