+++ .pre-commit-config.yaml
@@ -12,7 +12,7 @@ repos:
   - id: isort
     name: isort (python)
 - repo: https://github.com/macisamuele/language-formatters-pre-commit-hooks
-  rev: v2.7.0
+  rev: v2.8.0
   hooks:
   - id: pretty-format-yaml
     args: [--autofix]
@@ -24,7 +24,7 @@ repos:
   hooks:
   - id: black
 - repo: https://github.com/charliermarsh/ruff-pre-commit
-  rev: v0.0.255
+  rev: v0.0.257
   hooks:
   - id: ruff
 - repo: https://github.com/pre-commit/mirrors-mypy
