+++ .pre-commit-config.yaml
@@ -20,11 +20,11 @@ repos:
     exclude: poetry.lock
     args: [--autofix]
 - repo: https://github.com/psf/black
-  rev: 23.1.0
+  rev: 23.3.0
   hooks:
   - id: black
 - repo: https://github.com/charliermarsh/ruff-pre-commit
-  rev: v0.0.257
+  rev: v0.0.260
   hooks:
   - id: ruff
 - repo: https://github.com/pre-commit/mirrors-mypy
