+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 4.11.0
+current_version = 4.12.0
 commit = True
 message = chore(version): bump {current_version} to {new_version}
 
