### 2023-04-26 #206 @whedon
[REVIEW]: Manim Slides: A Python package for presenting Manim content anywhere
**Submitting author:** @jeertmans (<a href="http://orcid.org/0000-0002-5579-5360">Jérome Eertmans</a>)
**Repository:** <a href="https://github.com/jeertmans/manim-slides" target ="_blank">https://github.com/jeertmans/manim-slides</a>
**Version:** v4.9.2
