+++ paper/paper.md
@@ -64,7 +64,7 @@ provide new features on a regular basis.
 
 # Easy to Use Commitment
 
-Manim Slides is commited to be an easy-to-use tool, when minimal installation
+Manim Slides is commited to be an easy-to-use tool, with minimal installation
 procedure and few modifications required. It can either be used locally with its
 graphical user interface (GUI), or shared via HTML thanks to the RevealJS
 Javascript package [@revealjs].
@@ -92,8 +92,8 @@ and posted it on YouTube.
 # Stability and releases
 
 Manim Slides is continously tested on most recent Python versions, both ManimCE
-and ManimGL, and on all major platforms: **Ubuntu**, **macOS** and **Windows**. As of Manim
-Slide's exposed API begin very minimal, and the variaty of tests that are
+and ManimGL, and on all major platforms: **Ubuntu**, **macOS** and **Windows**. Due to Manim
+Slide's exposed API being very minimal, and the variety of tests that are
 performed, this tool can be considered stable over time.
 
 New releases are very frequent, as they mostly introduce enhancements or small
@@ -116,7 +116,7 @@ how to actually present them efficiently.
 We acknowledge the work of [@manim-presentation] that paved the initial structure
 of Manim Slides with the manim-presentation Python package.
 
-We also acknowledge Grant Sanderson for its termendous work on Manim, as well as
+We also acknowledge Grant Sanderson for his tremendous work on Manim, as
 well as the Manim Community contributors.
 
 Finally, we also acknowledge contributions from the GitHub contributors on the
