### 2023-04-29 9c1fdbf
A  few suggestions for the documentation
