### 2023-04-29 jeertmans/manim-slides#169 @bryanwweber
DOC: Improve installation documentation with a full end-to-end procedure
### Describe the Issue

From the JOSE review criteria:

>  Installation: Does installation proceed as outlined in the documentation? (and documentation is sufficient?)

The installation documentation clearly states the need to install ManimCE or ManimGL, which is appreciated. I'm curious why these are not declared as dependencies of the package? Especially in the quickstart, I think it would be nice to be opinionated about installing one or the other to show a full installation flow. It can always include a link to the other, but full instructions (I have only a venv created all the way to running an example) would be nice in the quickstart.

### Affected Page

https://eertmans.be/manim-slides/quickstart.html

### Issue Type

Documentation Enhancement

### Recommended fix or suggestions

_No response_
