### 2023-04-30 jeertmans/manim-slides#170 @bryanwweber
[FEATURE] Add wheel to PyPI and declare dependencies on Manim/ManimGL
### Description

From the JOSE review criteria:

>  Installation instructions: Is there a clearly stated list of dependencies? (Ideally these should be handled with an automated package management solution.)

I think it would be nice to improve the installation experience by declaring the dependency of Manim-slides on ManimCE/ManimGL so that these are installed automatically and with a supported version. Since you want to support both upstream packages, I think you could use an "extra" dependency, similar to your existing dev dependencies. Then the installation instructions could be modified to be `python -m pip install manim-slides[manim]` or `python -m pip install manim-slides[manimgl]`.

In addition, I think you should provide a wheel package on PyPI. Providing a wheel saves installation time for the user and they are not required to build anything locally when installing. I see you're using `build` but specifying only `--sdist`. Is there a particular reason to avoid the wheel files? Especially since this package is pure Python, it will only be one wheel uploaded.

I also suggest that you specify the compatible `python_requires` (or the equivalent poetry setting) so that Pip can determine whether the Python in the venv is supported. I suggest setting this as a lower bound and not an upper bound so that new Python releases are not limited in the version they can install, unless you know specifically that a particular version is not supported.

### Screenshots

_No response_

### Additional information

_No response_
