### 2023-04-30 jeertmans/manim-slides#171 @bryanwweber
DOC: JOSE Paper: Improve statement of need
### Describe the Issue

From the JOSE Reviewer Guidelines:

> A statement of need: Do the authors clearly state the need for this software and who the target audience is?

The paper.md does include a brief statement of need for the software, explaining that being able create slideshows from Manim animations is useful. I agree with this. I think the authors should also include information about the target audience and ways that those audiences might find this software useful, specifically. This does not have to be long, maybe 1-2 paragraphs at most.

### Affected Page

paper.md

### Issue Type

Documentation Enhancement

### Recommended fix or suggestions

_No response_
