### 2023-04-30 jeertmans/manim-slides#172 @bryanwweber
DOC: Improve contributions documentation
### Describe the Issue

From the JOSE Reviewer Guidelines:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

The documentation does include several pages with contribution instructions. However, the second and third items are not covered. I'd suggest adding links and explanations of these two items in the contributing instructions and on the README.

### Affected Page

https://eertmans.be/manim-slides/contributing/index.html

### Issue Type

Documentation Enhancement

### Recommended fix or suggestions

_No response_
