### 2023-04-30 #206 @bryanwweber
@jeertmans Thanks for the paper and the package! I've created several issues in the repository related to my review:

* jeertmans/manim-slides#168
* jeertmans/manim-slides#169
* jeertmans/manim-slides#170
* jeertmans/manim-slides#171
* jeertmans/manim-slides#172

Aside from those issues, the version that is tagged here for the paper is not the most current release. @magsol How should that version be updated? Thank you! I look forward to the responses.
