### 2023-04-30 #206 @jeertmans
Thanks for your review @bryanwweber! Indeed, the package has evolved quite a bit since the pre-review process was started. Most importantly, I have added a feature that converts Manim Slides presentations into PowerPoint files, which is pretty convenient for presenting in conferences for example.

I did not want to update the paper prior to the first review, but I can’t surely update the paper so it matches the latest version of Manim Slides.

For the rest of your review, I will address the comments directly in the PRs.
