### 2023-05-01 f46e323
chore(deps): add manim and manimgl as extras
