+++ pyproject.toml
@@ -49,6 +49,8 @@ version = "4.12.0"
 click = "^8.1.3"
 click-default-group = "^1.2.2"
 lxml = "^4.9.2"
+manim = { version = "^0.17.0", optional = true }
+manimgl = { version = "^1.6.1", optional = true }
 numpy = "^1.19"
 opencv-python = "^4.6.0.66"
 pydantic = "^1.10.2"
@@ -58,6 +60,10 @@ python-pptx = "^0.6.21"
 requests = "^2.28.1"
 rich = "^13.3.2"
 tqdm = "^4.64.1"
+
+[tool.poetry.extras]
+manim = ["manim"]
+manimgl = ["manimgl"]
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.10.0"
