### 2023-05-01 f41bc1e
chore(ci): run tests on dep. changes
