### 2023-05-01 0e82e28
chore(ci): add action dependabot checks (#174)
Setup dependabot to check for new actions
