### 2023-05-01 ce7e624
chore(docs): update install documentation
Closes #169
