### 2023-05-01 60f88be
chore(docs): update contributing guidelines
Add seeking for help and reporting an issue sections.

Closes #172
