### 2023-05-01 #206 @jeertmans
Hello @bryanwweber, I have opened a series of PRs to address each of your comments.
I hope they answer your concerns correctly, and I hope open to any new suggestion you might have :-)

Already, thank you for your time!
