### 2023-05-03 b05b7c4
feat(cli): add more debugging messages
