+++ manim_slides/logger.py
@@ -6,6 +6,8 @@ https://github.com/ManimCommunity/manim/blob/d5b65b844b8ce8ff5151a2f56f9dc98cebb
 import logging
 
 from rich.logging import RichHandler
+from rich.console import Console
+from rich.theme import Theme
 
 __all__ = ["logger", "make_logger"]
 
@@ -33,9 +35,12 @@ def make_logger() -> logging.Logger:
     RichHandler.KEYWORDS = HIGHLIGHTED_KEYWORDS
     rich_handler = RichHandler(
         show_time=True,
+        console=Console(theme=Theme({"logging.level.perf": "magenta"})),
     )
+    logging.addLevelName(5, "PERF")
     logger = logging.getLogger("manim-slides")
     logger.addHandler(rich_handler)
+
     return logger
 
 
