### 2023-05-03 536811f
chore(ci): udpate version in __version__ too
