+++ manim_slides/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "4.13.0-alpha"
+__version__ = "4.12.0"
