### 2023-05-07 d6ad561
chore(docs): update contributing guidelines (#177)
Add seeking for help and reporting an issue sections.

Closes #172
