### 2023-05-07 #206 @jeertmans
Thanks @bryanwweber! I've merged all the PRs and released the changes under v4.13.0, which also contains more debugging messages (explaining the bump in minor version).

Tell me if you need anything else to be done :-)
