### 2023-05-08 1ab9d1e
fix(convert): `--use-template` fixed
As described in #181, there was a mismatch between the type return by `click` and the one used by `pydantic`. Now we only use `Path` types.

Closes #181
