+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 4.13.0
+current_version = 4.13.1
 commit = True
 message = chore(version): bump {current_version} to {new_version}
 
