### 2023-05-08 396fa7a
fix(pages): missing assets
This is a PR to try understanding why some assets are not present
