### 2023-05-08 bb3e0ba
fix(ci): correct path
