+++ .github/workflows/pages.yml
@@ -48,12 +48,6 @@ jobs:
         sudo apt-get install libcairo2-dev libpango1.0-dev ffmpeg freeglut3-dev
     - name: Install local Python package
       run: poetry install --extras=manim --with docs
-    - name: Restore cached media
-      id: cache-media-restore
-      uses: actions/cache/restore@v3
-      with:
-        path: media
-        key: ${{ runner.os }}-media
     - name: Build animations
       run: |
         poetry run manim example.py ConvertExample BasicExample ThreeDExample
@@ -64,18 +58,6 @@ jobs:
         poetry run manim-slides convert -v DEBUG ThreeDExample docs/source/_static/three_d_example.html -ccontrols=true
     - name: Show docs/source/_static/ dir content (video only)
       run: tree -L 3 docs/source/_static/ -P '*.mp4'
-    - name: Clear cache
-      run: |
-        gh extension install actions/gh-actions-cache
-        gh actions-cache delete ${{ steps.cache-media-restore.outputs.cache-primary-key }} --confirm || true
-      env:
-        GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-    - name: Save media to cache
-      id: cache-media-save
-      uses: actions/cache/save@v3
-      with:
-        path: media
-        key: ${{ steps.cache-media-restore.outputs.cache-primary-key }}
     - name: Build docs
       run: cd docs && poetry run make html
     - name: Upload artifact
