+++ .pre-commit-config.yaml
@@ -24,7 +24,7 @@ repos:
   hooks:
   - id: black
 - repo: https://github.com/charliermarsh/ruff-pre-commit
-  rev: v0.0.263
+  rev: v0.0.265
   hooks:
   - id: ruff
 - repo: https://github.com/pre-commit/mirrors-mypy
