### 2023-05-11 90dd1a7
chore(docs): document scene subclassing
