### 2023-05-22 jeertmans/manim-slides#188 @behollister
DOC: JOSE Paper (paper.md) Feature Improvements Over Forked Repo
### Describe the Issue

Related to JOSE Review https://github.com/openjournals/jose-reviews/issues/206

paper.md does not include enumerated list of features added or changed over the original forked repo https://github.com/galatolofederico/manim-presentation

While a perusal of the  submissions repo itself can clarify it, it should be summarized more specifically in the paper.md for the submission.

### Affected Page

https://github.com/openjournals/jose-papers/blob/jose.00206/jose.00206/10.21105.jose.00206.pdf

### Issue Type

Documentation Enhancement

### Recommended fix or suggestions

include enumerated list of features added or changed over the original forked repo
