### 2023-05-22 #206 @behollister
@jeertmans @magsol added issue related to review

https://github.com/jeertmans/manim-slides/issues/188
