### 2023-05-23 c655b32
chore(paper): add comparison section
Closes #188
