### 2023-05-23 jeertmans/manim-slides#188 @behollister
PR satisfies issue concern per https://github.com/openjournals/jose-reviews/issues/206
