### 2023-05-23 jeertmans/manim-slides#194 @behollister
[FEATURE] Installation of manim and/or manimgl via manim-slides 
### Description

currently either manim or manimgl need to be installed separately and prior to manim-slides.

each, in turn, have their own set of dependencies.

best if manim-slides handles installing manim (or manimgl) as well, if not already present on target system.

request related to jose review: https://github.com/openjournals/jose-reviews/issues/206

### Screenshots

_No response_

### Additional information

_No response_
