### 2023-05-23 #206 @behollister
@jeertmans @magsol included additional issue related to review https://github.com/jeertmans/manim-slides/issues/194
