### 2023-05-24 jeertmans/manim-slides#194 @jeertmans
Hello @behollister!

This was already addressed in #170 and approved #173 as a solution (see GitHub approve).

Is it ok for you?
