### 2023-05-24 a1c041d
chore(paper): add comparison section (#192)
Closes #188
