### 2023-05-24 jeertmans/manim-slides#194 @behollister
Yes, all of the issues have been resolved.

Regards,
Brad

Brad E. Hollister, Ph.D.

Department of Computer Science
Californian State University Dominguez Hills
Office: NSM B218
Phone: 864-395-5560
