### 2023-05-25 #206 @jeertmans
Hello @magsol, thanks for recalling about the DOI! There you are: https://zenodo.org/record/7971361.

Tell me if I need to do anything else :-)
