### 2023-05-25 #206 @whedon
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
