### 2023-06-11 1e75cf9
feat(convert): add to PDF conversion
Basic PDF conversion. It takes the last frame (by default) for each animation, and prints out a PDF page.

Closes #196
