+++ pyproject.toml
@@ -53,6 +53,7 @@ manim = {version = "^0.17.0", optional = true}
 manimgl = {version = "^1.6.1", optional = true}
 numpy = "^1.19"
 opencv-python = "^4.6.0.66"
+pillow = "^9.5.0"
 pydantic = "^1.10.2"
 pyside6 = "^6.5.1.1"
 python = ">=3.8.1,<3.12"
