### 2023-06-11 b9f8f6a
chore(ci): remove experimental installer
