+++ .github/workflows/test_examples.yml
@@ -99,7 +99,6 @@ jobs:
       # Install Manim Slides
     - name: Install Manim Slides
       run: |
-        poetry config experimental.new-installer false
         poetry install --with test
 
     # Render slides
