+++ pyproject.toml
@@ -54,7 +54,7 @@ manimgl = {version = "^1.6.1", optional = true}
 numpy = "^1.19"
 opencv-python = "^4.6.0.66"
 pydantic = "^1.10.2"
-pyside6 = "^6.4.1"
+pyside6 = "^6.5.1.1"
 python = ">=3.8.1,<3.12"
 python-pptx = "^0.6.21"
 requests = "^2.28.1"
