+++ manim_slides/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "4.13.2"
+__version__ = "4.14.0"
