+++ .github/workflows/languagetool.yml
@@ -8,7 +8,7 @@ jobs:
   languagetool_check:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v1
+    - uses: actions/checkout@v3
     - uses: reviewdog/action-languagetool@v1
       with:
         reporter: github-pr-review
