+++ .pre-commit-config.yaml
@@ -23,8 +23,8 @@ repos:
   rev: 23.3.0
   hooks:
   - id: black
-- repo: https://github.com/charliermarsh/ruff-pre-commit
-  rev: v0.0.275
+- repo: https://github.com/astral-sh/ruff-pre-commit
+  rev: v0.0.276
   hooks:
   - id: ruff
 - repo: https://github.com/pre-commit/mirrors-mypy
