### 2023-07-05 f0c14b6
chore(deps): bumping pydantic to V2
