+++ pyproject.toml
@@ -54,13 +54,14 @@ manimgl = {version = "^1.6.1", optional = true}
 numpy = "^1.19"
 opencv-python = "^4.6.0.66"
 pillow = "^9.5.0"
-pydantic = "^1.10.2"
+pydantic = "^2.0.1"
 pyside6 = "^6.5.1.1"
 python = ">=3.8.1,<3.12"
 python-pptx = "^0.6.21"
 requests = "^2.28.1"
 rich = "^13.3.2"
 tqdm = "^4.64.1"
+pydantic-extra-types = "^2.0.0"
 
 [tool.poetry.extras]
 manim = ["manim"]
