### 2023-07-05 be5e222
fix(ci): add tests and fixes
