+++ tests/test_config.py
@@ -1,0 +1,45 @@
+import random
+import string
+from pathlib import Path
+from typing import Any, Generator, List
+
+import pytest
+from pydantic import ValidationError
+
+from manim_slides.config import Key, merge_basenames
+
+
+def random_path(
+    length: int = 20,
+    dirname: Path = Path("./media/videos/example"),
+    suffix: str = ".mp4",
+) -> Path:
+    basename = "".join(random.choices(string.ascii_letters, k=length))
+
+    return dirname.joinpath(basename + suffix)
+
+
+@pytest.fixture
+def paths() -> Generator[List[Path], None, None]:
+    random.seed(1234)
+
+    yield [random_path() for _ in range(20)]
+
+
+def test_merge_basenames(paths: List[Path]) -> None:
+    path = merge_basenames(paths)
+    assert path.suffix == paths[0].suffix
+    assert path.parent == paths[0].parent
+
+
+class TestKey:
+    @pytest.mark.parametrize(("ids", "name"), [([1], None), ([1], "some key name")])
+    def test_valid_keys(self, ids: Any, name: Any) -> None:
+        _ = Key(ids=ids, name=name)
+
+    @pytest.mark.parametrize(
+        ("ids", "name"), [([]), ([-1], None), ([1], {"an": " invalid name"})]
+    )
+    def test_invalid_keys(self, ids: Any, name: Any) -> None:
+        with pytest.raises(ValidationError):
+            _ = Key(ids=ids, name=name)
