### 2023-07-05 fe0912c
fix(lib): add missing mode arg
