### 2023-07-05 c38cf0f
chore(tests): add more tests and use pytest
