### 2023-07-05 17cefad
chore(ci): install manim deps
