### 2023-07-05 8fa6644
fix(lib): add custom schema
