+++ pyproject.toml
@@ -86,6 +86,7 @@ sphinxext-opengraph = "^0.7.5"
 
 [tool.poetry.group.test.dependencies]
 pytest = "^7.4.0"
+pytest-cov = "^4.1.0"
 
 [tool.poetry.plugins]
 
