### 2023-07-06 #206 @magsol
Hi @openjournals/jose-eics, just wanted to ping you on this one being ready for publication.
