### 2023-07-06 60f284e
Update index.md (#210)
Small typo
