+++ docs/source/index.md
@@ -26,7 +26,7 @@ Manim Slides makes creating slides with Manim super easy!
 In a [very few steps](./quickstart), you can create slides and present them either using the GUI, or your browser.
 
 
-Slide through the demo below to get a quick glimpse on what you can do with Manin Slides.
+Slide through the demo below to get a quick glimpse on what you can do with Manim Slides.
 
 
 <!-- From: https://faq.dailymotion.com/hc/en-us/articles/360022841393-How-to-preserve-the-player-aspect-ratio-on-a-responsive-page -->
