### 2023-07-07 a353de2
chore(docs): fix typo
