### 2023-07-19 4291fa3
chore(ci): apply ruff's suggestion
