### 2023-07-19 8c3c3a3
feat(lib/docs): add sphinx directive
This PR adds a new Sphinx directive, inspired by the Manim one. This directive is also used within the docs to display examples.
