+++ docs/source/reference/sphinx_extension.md
@@ -1,0 +1,6 @@
+# Manim Slides' Sphinx directive
+
+```{eval-rst}
+.. automodule:: manim_slides.docs.manim_slides_directive
+   :members:
+```
