+++ pyproject.toml
@@ -76,10 +76,10 @@ pre-commit = "^3.0.2"
 ruff = "^0.0.219"
 
 [tool.poetry.group.docs.dependencies]
-furo = "^2022.9.29"
+furo = "^2023.5.20"
 manim = "^0.17.0"
-myst-parser = "^0.18.1"
-sphinx = "^5.3.0"
+myst-parser = "^2.0.0"
+sphinx = "^7.0.1"
 sphinx-click = "^4.4.0"
 sphinx-copybutton = "^0.5.1"
 sphinxext-opengraph = "^0.7.5"
