### 2023-07-20 66629a5
feat(lib): improve wipe animation
Improve wipe transition by using FadeIn and FadeOut. The also allows to support introducers and removers.
