+++ manim_slides/manim.py
@@ -18,6 +18,8 @@ __all__ = [
     "MANIMGL_IMPORTED",
     # Classes
     "AnimationGroup",
+    "FadeIn",
+    "FadeOut",
     "Mobject",
     "Scene",
     "ThreeDScene",
@@ -73,7 +75,7 @@ else:
 
 
 if MANIMGL:
-    from manimlib import LEFT, AnimationGroup, Mobject, Scene, ThreeDScene, config
+    from manimlib import LEFT, AnimationGroup, Mobject, Scene, ThreeDScene, config, FadeIn, FadeOut
     from manimlib.constants import FFMPEG_BIN
     from manimlib.logger import log as logger
 
@@ -87,6 +89,8 @@ else:
             ThreeDScene,
             config,
             logger,
+            FadeIn,
+            FadeOut
         )
 
         try:  # For manim<v0.16.0.post0
