+++ manim_slides/manim.py
@@ -18,6 +18,8 @@ __all__ = [
     "MANIMGL_IMPORTED",
     # Classes
     "AnimationGroup",
+    "FadeIn",
+    "FadeOut",
     "Mobject",
     "Scene",
     "ThreeDScene",
@@ -73,7 +75,16 @@ else:
 
 
 if MANIMGL:
-    from manimlib import LEFT, AnimationGroup, Mobject, Scene, ThreeDScene, config
+    from manimlib import (
+        LEFT,
+        AnimationGroup,
+        FadeIn,
+        FadeOut,
+        Mobject,
+        Scene,
+        ThreeDScene,
+        config,
+    )
     from manimlib.constants import FFMPEG_BIN
     from manimlib.logger import log as logger
 
@@ -82,6 +93,8 @@ else:
         from manim import (
             LEFT,
             AnimationGroup,
+            FadeIn,
+            FadeOut,
             Mobject,
             Scene,
             ThreeDScene,
