### 2023-07-20 a10902e
feat(lib): add `wait-time-between-slides` attribute (#218)
Add a new attribute that will allow to automatically call `self.wait(...)` between two slides, to fix graphical issues.
