### 2023-07-20 60c3f57
feat(lib): add canvas and features
Add canvas to slide
