### 2023-07-20 b195b82
chore(docs): improve directive's documentation
