### 2023-07-20 66a63df
chore(ci): same but for code coverage
