+++ .github/workflows/coverage.yml
@@ -6,6 +6,11 @@ jobs:
   test:
     name: Coverage
     runs-on: ubuntu-latest
+    env:
+      QT_QPA_PLATFORM: offscreen
+      MANIM_SLIDES_VERBOSITY: debug
+      PYTHONFAULTHANDLER: 1
+      DISPLAY: :99
     steps:
     - name: Checkout repository
       uses: actions/checkout@v3
@@ -22,6 +27,11 @@ jobs:
     - name: Install manim dependencies on Ubuntu
       run: |
         sudo apt-get install libcairo2-dev libpango1.0-dev ffmpeg freeglut3-dev
+
+    - name: Install xvfb on Ubuntu
+      run: |
+        sudo apt-get install xvfb
+        nohup Xvfb $DISPLAY &
 
     - name: Install Manim Slides
       run: |
