+++ .github/workflows/tests.yml
@@ -11,6 +11,11 @@ jobs:
       matrix:
         pyversion: ['3.8', '3.9', '3.10', '3.11']
     runs-on: ubuntu-latest
+    env:
+      QT_QPA_PLATFORM: offscreen
+      MANIM_SLIDES_VERBOSITY: debug
+      PYTHONFAULTHANDLER: 1
+      DISPLAY: :99
     steps:
     - name: Checkout repository
       uses: actions/checkout@v3
@@ -27,6 +32,11 @@ jobs:
     - name: Install manim dependencies on Ubuntu
       run: |
         sudo apt-get install libcairo2-dev libpango1.0-dev ffmpeg freeglut3-dev
+
+    - name: Install xvfb on Ubuntu
+      run: |
+        sudo apt-get install xvfb
+        nohup Xvfb $DISPLAY &
 
     - name: Install Manim Slides
       run: |
