+++ pyproject.toml
@@ -85,6 +85,8 @@ sphinx-copybutton = "^0.5.1"
 sphinxext-opengraph = "^0.7.5"
 
 [tool.poetry.group.test.dependencies]
+manim = "^0.17.3"
+manimgl = "^1.6.1"
 pytest = "^7.4.0"
 pytest-cov = "^4.1.0"
 
@@ -92,6 +94,12 @@ pytest-cov = "^4.1.0"
 
 [tool.poetry.plugins."console_scripts"]
 manim-slides = "manim_slides.__main__:cli"
+
+[tool.pytest.ini_options]
+filterwarnings = [
+  "error",
+  "ignore::DeprecationWarning"
+]
 
 [tool.ruff]
 ignore = [
