### 2023-07-20 8069ab5
chore(docs): improve example animation
