### 2023-07-24 f7c320d
fix(lib/cli): fix validation error
