### 2023-07-24 c9a49f4
chore(tests): test commons and defaults
Add new tests to improve overall coverage
