+++ tests/conftest.py
@@ -1,0 +1,3 @@
+from manim_slides.logger import make_logger
+
+_ = make_logger()  # This is run so that "PERF" level is created
