### 2023-07-24 2b6240c
chore(tests): test commons and defaults (#222)
Add new tests to improve overall coverage
