### 2023-07-24 fdb2a9b
fix(lib/cli): relative paths, empty slides, and tests
This fixes two issues:

1. Empty slides are now reported as error, to prevent indexing error;
2. Changing the folder path will now produce an absolute path to slides, which was not the case before and would lead to a "file does not exist error".

A few tests were also added to cover those
