+++ pyproject.toml
@@ -61,6 +61,7 @@ python = ">=3.8.1,<3.12"
 python-pptx = "^0.6.21"
 requests = "^2.28.1"
 rich = "^13.3.2"
+rtoml = "^0.9.0"
 tqdm = "^4.64.1"
 
 [tool.poetry.extras]
