+++ manim_slides/defaults.py
@@ -1,3 +1,3 @@
 FOLDER_PATH: str = "./slides"
-CONFIG_PATH: str = ".manim-slides.json"
+CONFIG_PATH: str = ".manim-slides.toml"
 FFMPEG_BIN: str = "ffmpeg"
