### 2023-07-24 6af5f49
chore(lib): remove all os.path in favor to pathlib
