+++ manim_slides/defaults.py
@@ -1,3 +1,5 @@
-FOLDER_PATH: str = "./slides"
-CONFIG_PATH: str = ".manim-slides.toml"
-FFMPEG_BIN: str = "ffmpeg"
+from pathlib import Path
+
+FOLDER_PATH: Path = Path("./slides")
+CONFIG_PATH: Path = Path(".manim-slides.toml")
+FFMPEG_BIN: Path = Path("ffmpeg")
