### 2023-07-24 8efbfc2
fix(lib): str to path
