+++ .pre-commit-config.yaml
@@ -24,7 +24,7 @@ repos:
   hooks:
   - id: black
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.0.280
+  rev: v0.0.281
   hooks:
   - id: ruff
 - repo: https://github.com/pre-commit/mirrors-mypy
