### 2023-08-01 aaa70d4
feat(lib): add zoom animation
