### 2023-08-02 01fe137
fix(lib): check if reversed file exist
This fixes a bug that would block the process forever because FFMPEG was asking if it could overwrite or not.
