### 2023-08-02 43ab0cb
chore(lib): fix zoom animation
