+++ manim_slides/slide.py
@@ -606,7 +606,7 @@ class Slide(Scene):  # type:ignore
                     text = Text("This is a wipe example").next_to(square, DOWN)
                     beautiful = Text("Beautiful, no?")
 
-                    self.play(Create(circle))
+                    self.play(FadeIn(circle))
                     self.next_slide()
 
                     self.play(self.wipe(circle, Group(square, text)))
@@ -628,6 +628,69 @@ class Slide(Scene):  # type:ignore
 
         for mobject in current:
             animations.append(FadeOut(mobject, shift=shift_amount, **fade_out_kwargs))
+
+        return AnimationGroup(*animations, **kwargs)
+
+    def zoom(
+        self,
+        current: Sequence[Mobject] = [],
+        future: Sequence[Mobject] = [],
+        scale: float = 4.0,
+        out: bool = False,
+        fade_in_kwargs: Mapping[str, Any] = {},
+        fade_out_kwargs: Mapping[str, Any] = {},
+        **kwargs: Any,
+    ) -> AnimationGroup:
+        """
+        Returns a zoom animation that will fade out all the current objects,
+        and fade in all the future objects. Objects are faded in a direction
+        that goes towards the camera.
+
+        :param current: A sequence of mobjects to remove from the scene.
+        :param future: A sequence of mobjects to add to the scene.
+        :param scale: How much the objects are scaled (up or down).
+        :param out: If set, the objects fade in the opposite direction.
+        :param fade_in_kwargs: Keyword arguments passed to
+            :class:`FadeIn<manim.animation.fading.FadeIn>`.
+        :param fade_out_kwargs: Keyword arguments passed to
+            :class:`FadeOut<manim.animation.fading.FadeOut>`.
+        :param kwargs: Keyword arguments passed to
+            :class:`AnimationGroup<manim.animation.composition.AnimationGroup>`.
+
+        Examples
+        --------
+
+        .. manim-slides:: ZoomExample
+
+            from manim import *
+            from manim_slides import Slide
+
+            class ZoomExample(Slide):
+                def construct(self):
+                    circle = Circle(radius=3, color=BLUE)
+                    square = Square()
+
+                    self.play(FadeIn(circle))
+                    self.next_slide()
+
+                    self.play(self.zoom(circle, square))
+                    self.next_slide()
+
+                    self.play(self.zoom(square, circle, out=True, scale=10.))
+        """
+        scale_in = 1.0 / scale
+        scale_out = scale
+
+        if out:
+            scale_in, scale_out = scale_out, scale_in
+
+        animations = []
+
+        for mobject in future:
+            animations.append(FadeIn(mobject, scale=scale_in, **fade_in_kwargs))
+
+        for mobject in current:
+            animations.append(FadeOut(mobject, scale=scale_out, **fade_out_kwargs))
 
         return AnimationGroup(*animations, **kwargs)
 
