+++ pyproject.toml
@@ -49,7 +49,7 @@ version = "4.14.0"
 click = "^8.1.3"
 click-default-group = "^1.2.2"
 lxml = "^4.9.2"
-manim = {version = "^0.17.0", optional = true}
+manim = {version = "^0.17.3", optional = true}
 manimgl = {version = "^1.6.1", optional = true}
 numpy = "^1.19"
 opencv-python = "^4.6.0.66"
@@ -78,7 +78,7 @@ ruff = "^0.0.219"
 
 [tool.poetry.group.docs.dependencies]
 furo = "^2023.5.20"
-manim = "^0.17.0"
+manim = "^0.17.3"
 myst-parser = "^2.0.0"
 sphinx = "^7.0.1"
 sphinx-click = "^4.4.0"
