### 2023-08-02 0de877e
fix(docs): typo in RST
