### 2023-08-04 fc36909
chore(tests): zoom example
