### 2023-08-04 f01f811
chore(present): inform about SIGINT
