### 2023-08-04 713b9e1
feat(present): fix presentation size and add screen option
This kinda fixes the issue of weird resizes between two presentations
