### 2023-08-04 9e1e0f2
fix(lib): correctly caching and simplified structure (#233)
