### 2023-08-04 dec2f5e
chore(docs): actually apply my own recomm. in example :-)
