+++ manim_slides/convert.py
@@ -4,6 +4,7 @@ import subprocess
 import sys
 import tempfile
 import webbrowser
+from base64 import b64encode
 from enum import Enum
 from importlib import resources
 from pathlib import Path
@@ -34,6 +35,29 @@ from .logger import logger
 from .present import get_scenes_presentation_config
 
 
+DATA_URI_FIX = r"""
+// Fix found by @t-fritsch on GitHub
+// see: https://github.com/hakimel/reveal.js/discussions/3362#discussioncomment-6651475.
+function fixBase64VideoBackground(event) {
+    // event.previousSlide, event.currentSlide, event.indexh, event.indexv
+    if (event.currentSlide.getAttribute('data-background-video')) {
+        const background = Reveal.getSlideBackground(event.indexh, event.indexv),
+            video = background.querySelector('video'),
+            sources = video.querySelectorAll('source');
+
+        sources.forEach((source, i) => {
+            const src = source.getAttribute('src');
+            if(src.match(/^data:video.*;base64$/)){
+                const nextSrc = sources[i+1]?.getAttribute('src');
+                video.setAttribute('src', `${src},${nextSrc}`);
+            }
+        });
+    }
+}
+Reveal.on( 'ready', fixBase64VideoBackground );
+Reveal.on( 'slidechanged', fixBase64VideoBackground );
+"""
+
 def open_with_default(file: Path) -> None:
     system = platform.system()
     if system == "Darwin":
@@ -59,6 +83,15 @@ def validate_config_option(
             )
 
     return config
+
+
+def data_uri(file: Path) -> str:
+    """
+    Reads a video and returns the corresponding data-uri.
+    """
+    b64 = b64encode(file.read_bytes()).decode()
+
+    return f"data:video/webm;base64,{b64}"
 
 
 class Converter(BaseModel):  # type: ignore
@@ -238,6 +271,8 @@ class RevealTheme(str, Enum):
 
 
 class RevealJS(Converter):
+    # Export option: use data-uri
+    data_uri: bool = False
     # Presentation size options from RevealJS
     width: Union[Str, int] = Str("100%")
     height: Union[Str, int] = Str("100%")
@@ -326,9 +361,13 @@ class RevealJS(Converter):
         for presentation_config in self.presentation_configs:
             for slide_config in presentation_config.slides:
                 file = presentation_config.files[slide_config.start_animation]
-                file = assets_dir / file.name
 
                 logger.debug(f"Writing video section with file {file}")
+
+                if self.data_uri:
+                    file = data_uri(file)
+                else:
+                    file = assets_dir / file.name
 
                 # TODO: document this
                 # Videos are muted because, otherwise, the first slide never plays correctly.
@@ -356,27 +395,36 @@ class RevealJS(Converter):
 
     def convert_to(self, dest: Path) -> None:
         """Converts this configuration into a RevealJS HTML presentation, saved to DEST."""
-        dirname = dest.parent
-        basename = dest.stem
-        ext = dest.suffix
+        if self.data_uri:
+            assets_dir = Path("")  # Actually we won't care.
+        else:
+            dirname = dest.parent
+            basename = dest.stem
+            ext = dest.suffix
 
-        assets_dir = Path(
-            self.assets_dir.format(dirname=dirname, basename=basename, ext=ext)
-        )
-        full_assets_dir = dirname / assets_dir
+            assets_dir = Path(
+                self.assets_dir.format(dirname=dirname, basename=basename, ext=ext)
+            )
+            full_assets_dir = dirname / assets_dir
 
-        logger.debug(f"Assets will be saved to: {full_assets_dir}")
+            logger.debug(f"Assets will be saved to: {full_assets_dir}")
 
-        os.makedirs(full_assets_dir, exist_ok=True)
[…]
