### 2023-08-06 790abfa
chore(lib): explicit decode type
