### 2023-08-06 d66bc61
feat(lib): auto detect mime type
