+++ pyproject.toml
@@ -90,6 +90,8 @@ manim = "^0.17.3"
 manimgl = "^1.6.1"
 pytest = "^7.4.0"
 pytest-cov = "^4.1.0"
+pytest-env = "^0.8.2"
+pytest-xdist = "^3.3.1"
 
 [tool.poetry.plugins]
 
@@ -97,6 +99,9 @@ pytest-cov = "^4.1.0"
 manim-slides = "manim_slides.__main__:cli"
 
 [tool.pytest.ini_options]
+env = [
+  "QT_QPA_PLATFORM=offscreen",
+]
 filterwarnings = [
   "error",
   "ignore::DeprecationWarning"
