### 2023-08-06 #206 @labarba
Hi everybody! 👋  — It's me who's backlogged due to travel, keynote, more travel, and email overload 😬
Thank you for your patience.

For some reason, the archive does not show on the top of the issue...

But I see that @magsol ran the command to add it [above](https://github.com/openjournals/jose-reviews/issues/206#issuecomment-1563258724).

I will try to add it again with our upgraded editorial bot...
