### 2023-08-06 #206 @labarba
have a look at this tiny fix: https://github.com/labarba/manim-slides/pull/1
