### 2023-08-06 146a2f7
chore(paper): fix brackets
As proposed by @labarba
