### 2023-08-06 #206 @jeertmans
Thanks for your suggestion @labarba, this is fixed :-)

As the v4.15.0 was released, and the paper is still not published, maybe it would be possible to update the version to v4.15.0, and the Zenodo to 10.5281/zenodo.8215167?
