### 2023-08-06 #206 @labarba
Ah. Do you have the GitHub Zenodo integration set up to update the archive automatically with each release? We do like to ask authors to edit the metadata of the archive so that the title and author list match the paper. (The auto-archives pull as authors all committers to the repo and use the repo name as title.) Could you do that change for the archive we will attach to the JOSE paper?
