### 2023-08-07 fc157b9
feat(lib): add Jupyter magic
And also use the same logger level as manim (by default)
