+++ docs/source/reference/index.md
@@ -10,6 +10,7 @@ cli
 examples
 gui
 html
+IPython magic <ipython_magic>
 sharing
 Sphinx Extension <sphinx_extension>
 ```
@@ -25,6 +26,8 @@ Slides' executable.
 [Graphical User Interface](./gui): details about the main Manim Slide' feature.
 
 [HTML Presentation](./html): an alternative way of presenting your animations.
+
+[IPython Magic](./ipython_magic): a magic to render and display Manim Slides inside notebooks.
 
 [Sharing](./sharing): how to share your presentation with others.
 
