+++ pyproject.toml
@@ -63,10 +63,15 @@ requests = "^2.28.1"
 rich = "^13.3.2"
 rtoml = "^0.9.0"
 tqdm = "^4.64.1"
+notebook = {version = "^7.0.2", optional = true}
+docutils = {version = "^0.20.1", optional = true}
+jinja2 = {version = "^3.1.2", optional = true}
 
 [tool.poetry.extras]
+magic = ["manim", "notebook"]
 manim = ["manim"]
 manimgl = ["manimgl"]
+sphinx-directive = ["docutils", "jinja2", "manim"]
 
 [tool.poetry.group.dev.dependencies]
 black = "^22.10.0"
