+++ pyproject.toml
@@ -48,9 +48,12 @@ version = "4.15.0"
 [tool.poetry.dependencies]
 click = "^8.1.3"
 click-default-group = "^1.2.2"
+docutils = {version = "^0.20.1", optional = true}
+jinja2 = {version = "^3.1.2", optional = true}
 lxml = "^4.9.2"
 manim = {version = "^0.17.3", optional = true}
 manimgl = {version = "^1.6.1", optional = true}
+notebook = {version = "^7.0.2", optional = true}
 numpy = "^1.19"
 opencv-python = "^4.6.0.66"
 pillow = "^9.5.0"
@@ -63,9 +66,6 @@ requests = "^2.28.1"
 rich = "^13.3.2"
 rtoml = "^0.9.0"
 tqdm = "^4.64.1"
-notebook = {version = "^7.0.2", optional = true}
-docutils = {version = "^0.20.1", optional = true}
-jinja2 = {version = "^3.1.2", optional = true}
 
 [tool.poetry.extras]
 magic = ["manim", "notebook"]
