### 2023-08-07 f65818a
fix(lib): remove deleted module
