### 2023-08-07 a563b35
chore(lib): fix typing issues
