### 2023-08-07 ff3c4ab
chore(docs): document magic
