+++ pyproject.toml
@@ -66,9 +66,10 @@ requests = "^2.28.1"
 rich = "^13.3.2"
 rtoml = "^0.9.0"
 tqdm = "^4.64.1"
+ipython = {version = ">=8.12.2", optional = true}
 
 [tool.poetry.extras]
-magic = ["manim", "notebook"]
+magic = ["manim", "ipython"]
 manim = ["manim"]
 manimgl = ["manimgl"]
 sphinx-directive = ["docutils", "jinja2", "manim"]
@@ -89,6 +90,8 @@ sphinx = "^7.0.1"
 sphinx-click = "^4.4.0"
 sphinx-copybutton = "^0.5.1"
 sphinxext-opengraph = "^0.7.5"
+nbsphinx = "^0.9.2"
+pandoc = "^2.3"
 
 [tool.poetry.group.test.dependencies]
 manim = "^0.17.3"
