+++ pyproject.toml
@@ -49,6 +49,7 @@ version = "4.15.0"
 click = "^8.1.3"
 click-default-group = "^1.2.2"
 docutils = {version = "^0.20.1", optional = true}
+ipython = {version = ">=8.12.2", optional = true}
 jinja2 = {version = "^3.1.2", optional = true}
 lxml = "^4.9.2"
 manim = {version = "^0.17.3", optional = true}
@@ -66,7 +67,6 @@ requests = "^2.28.1"
 rich = "^13.3.2"
 rtoml = "^0.9.0"
 tqdm = "^4.64.1"
-ipython = {version = ">=8.12.2", optional = true}
 
 [tool.poetry.extras]
 magic = ["manim", "ipython"]
@@ -86,12 +86,12 @@ ruff = "^0.0.219"
 furo = "^2023.5.20"
 manim = "^0.17.3"
 myst-parser = "^2.0.0"
+nbsphinx = "^0.9.2"
+pandoc = "^2.3"
 sphinx = "^7.0.1"
 sphinx-click = "^4.4.0"
 sphinx-copybutton = "^0.5.1"
 sphinxext-opengraph = "^0.7.5"
-nbsphinx = "^0.9.2"
-pandoc = "^2.3"
 
 [tool.poetry.group.test.dependencies]
 manim = "^0.17.3"
