### 2023-08-07 7f9c73a
fix(ci): install kernel
