### 2023-08-07 cb8a79f
fix(ci): spawning is not necessary (and fails)
