### 2023-08-07 6ff6247
chore(ci): add ipykernel
