+++ pyproject.toml
@@ -84,6 +84,7 @@ ruff = "^0.0.219"
 
 [tool.poetry.group.docs.dependencies]
 furo = "^2023.5.20"
+ipykernel = "^6.25.1"
 manim = "^0.17.3"
 myst-parser = "^2.0.0"
 nbsphinx = "^0.9.2"
