+++ manim_slides/manim.py
@@ -1,8 +1,6 @@
 import sys
 from importlib.util import find_spec
 
-from .utils import suppress_stdout
-
 __all__ = [
     # Constants
     "FFMPEG_BIN",
@@ -77,20 +75,19 @@ if MANIMGL:
     from manimlib.logger import log as logger
 
 else:
-    with suppress_stdout():  # Avoids printing "Manim Community v..."
-        from manim import (
-            LEFT,
-            AnimationGroup,
-            FadeIn,
-            FadeOut,
-            Mobject,
-            Scene,
-            ThreeDScene,
-            config,
-            logger,
-        )
+    from manim import (
+        LEFT,
+        AnimationGroup,
+        FadeIn,
+        FadeOut,
+        Mobject,
+        Scene,
+        ThreeDScene,
+        config,
+        logger,
+    )
 
-        try:  # For manim<v0.16.0.post0
-            from manim.constants import FFMPEG_BIN
-        except ImportError:
-            FFMPEG_BIN = config.ffmpeg_executable
+    try:  # For manim<v0.16.0.post0
+        from manim.constants import FFMPEG_BIN
+    except ImportError:
+        FFMPEG_BIN = config.ffmpeg_executable
