+++ docs/source/reference/index.md
@@ -10,6 +10,7 @@ cli
 examples
 gui
 html
+IPython magic <ipython_magic>
 sharing
 Sphinx Extension <sphinx_extension>
 ```
@@ -26,6 +27,11 @@ Slides' executable.
 
 [HTML Presentation](./html): an alternative way of presenting your animations.
 
+[IPython Magic](./ipython_magic): a magic to render and display Manim Slides inside notebooks.
+
++ [Example](./magic_example): example notebook using the magics.
+
 [Sharing](./sharing): how to share your presentation with others.
+
 
 [Sphinx Extension](./sphinx_extension): a Sphinx extension for diplaying Manim Slides animations within your documentation.
