### 2023-08-07 #206 @labarba
@openjournals/dev 👋
This is our first time accepting a paper here at JOSE with the new editorial workflow and bot, and I'm afraid something has gone awry. Help!
