### 2023-08-08 #206 @jeertmans
@labarba That's weird, even more because I received a ORCID notification asking me the permission to link to that paper... which links to 404 not found https://joss.theoj.org/papers/10.21105/jose.00206

Note: I don't think the @openjournals/dev tag works
