### 2023-08-08 efc3017
chore(paper): add paper's doi, update badges and add CFF
