+++ .bumpversion.cfg
@@ -10,3 +10,7 @@ replace = __version__ = "{new_version}"
 [bumpversion:file:pyproject.toml]
 search = version = "{current_version}"
 replace = version = "{new_version}"
+
+[bumpversion:file:CITATION.cff]
+search = version: {current_version}
+replace = version: {new_version}
