### 2023-08-08 #206 @labarba
Congratulations, @jeertmans, your JOSE paper is published! 🎉

Huge thanks to. our Editor: @magsol and the Reviewers: @behollister, @bryanwweber — this is all possible thanks to your generous contributions 🙏
