### 2023-08-08 #206 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
