+++ .bumpversion.cfg
@@ -12,5 +12,5 @@ search = version = "{current_version}"
 replace = version = "{new_version}"
 
 [bumpversion:file:CITATION.cff]
-search = version: {current_version}
-replace = version: {new_version}
+search = version: v{current_version}
+replace = version: v{new_version}
