+++ CITATION.cff
@@ -29,4 +29,4 @@ keywords:
   - PowerPoint
   - Python
 license: MIT
-version: v4.15.0
+version: v4.16.0
