+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 4.15.0
+current_version = 4.16.0
 commit = True
 message = chore(version): bump {current_version} to {new_version}
 
